requirements.txt
alembic==0.9.3.dev0
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
Babel==2.4.0
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0.8
django-crispy-forms==1.8.1
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==1.7.0
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
language-selector==0.1
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
ldap3==2.4.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==1.0
netifaces==0.10.4
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
paramiko==2.0.0
passlib==1.7.1
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.4
psycopg2-binary==2.8.5
pyasn1==0.4.2
pycairo==1.16.2
pycups==1.9.73
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
speaklater==1.4
SQLAlchemy==1.1.11
sqlparse==0.3.1
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.16.0
whitenoise==5.0.1
WTForms==2.1
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
